Aruco working in linux with docstring and text
